Refactored entire application structure
