+++ package.json
@@ -4,8 +4,8 @@
   "main": "scripts.js",
   "scripts": {
     "start": "npm-run-all -p nodemon browsersync",
-    "nodemon": "nodemon server.js",
-    "browsersync": "browser-sync start --proxy http://localhost:5000 --files 'public,views'"
+    "nodemon": "nodemon src/server.js",
+    "browsersync": "browser-sync start --proxy http://localhost:5000 --files 'public,src/app/views'"
   },
   "dependencies": {
     "express": "^4.17.1",
