+++ public/styles.css
@@ -249,7 +249,7 @@ footer p {
 }
 
 
-/*=== ADMIN ===*/
+/*=== ADMIN HEADER ===*/
 
 header.admin {
     background-color: black;
@@ -266,6 +266,33 @@ header.admin {
     justify-content: flex-start;
     align-items: center;
 }
+
+header.admin .menu {
+    margin-left: 50px;
+}
+
+header.admin .menu a {
+    color: #999;
+    text-decoration: none;
+    font-weight: bold;
+    text-transform: uppercase;
+    font-size: 15px;
+    margin-right: 30px;
+}
+
+header.admin .menu a:hover {
+    filter: brightness(2);
+}
+
+header.admin .menu a.active {
+    color: white;
+}
+
+header.admin .menu a:last-child {
+    margin: 0;
+}
+
+/*=== ADMIN PAGE ===*/
 
 section#topRecipes div#topRecipesGrid.admin {
     margin-top: 30px;
