+++ src/app/controllers/admin.js
@@ -1,5 +1,5 @@
 const fs = require('fs')
-const data = require('../data.json')
+const data = require('../../../data.json')
 
 exports.index = (req, res) => {
     return res.render("admin/index", { recipes: data.recipes })
