+++ src/app/views/admin/layout.njk
@@ -11,6 +11,10 @@
     <header class="admin">
         <div class="headerContainer">
             <a href="/admin/recipes"><img src="/assets/logo_white.png" alt="Foodfy Logo"></a>
+            <div class="menu">
+                <a href="/admin/recipes">Receitas</a>
+                <a href="/admin/chefs">Chefs</a>
+            </div>
         </div>
     </header>
