+++ src/app/views/admin/recipe_info.njk
